# macrocomm/src/app/api/generated/model/incidentResponse.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IncidentSeverity } from './incidentSeverity';
import { IncidentStatus } from './incidentStatus';


export interface IncidentResponse { 
    id: number;
    vehicle_registration: string;
    location: string;
    issue_type: string;
    description: string;
    severity: IncidentSeverity;
    status: IncidentStatus;
    priority_score: number;
    created_at: string;
}
export namespace IncidentResponse {
}