# macrocomm/src/app/api/generated/model/incidentStatus.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const IncidentStatus = {
    New: 'new',
    Assigned: 'assigned',
    InProgress: 'in_progress',
    Resolved: 'resolved'
} as const;
export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];