# macrocomm/src/app/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Get Apihealth",
        "operationId": "get_apihealth_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    },
    "/api/incidents": {
      "get": {
        "summary": "Get Incidents",
        "operationId": "get_incidents_api_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response Get Incidents Api Incidents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Incident",
        "operationId": "create_incident_api_incidents_post",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/status": {
      "patch": {
        "summary": "Update Incident Status",
        "operationId": "update_incident_status_api_incidents__incident_id__status_patch",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Incident Id"
            }
          },
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response Update Incident Status Api Incidents  Incident Id  Status Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/summary": {
      "get": {
        "summary": "Get Incident Summary",
        "operationId": "get_incident_summary_api_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vehicle_registration": {
            "type": "string",
            "title": "Vehicle Registration"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "severity": {
            "$ref": "#/components/schemas/IncidentSeverity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "priority_score": {
            "type": "integer",
            "title": "Priority Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle_registration",
          "location",
          "issue_type",
          "description",
          "severity",
          "status",
          "priority_score",
          "created_at"
        ],
        "title": "IncidentResponse"
      },
      "IncidentSeverity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "IncidentSeverity"
      },
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "new",
          "assigned",
          "in_progress",
          "resolved"
        ],
        "title": "IncidentStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}